+++ satella/coding/structures/__init__.py
@@ -5,10 +5,12 @@ __all__ = [
     'OmniHashableMixin',
     'Singleton',
     'DictObject',
-    'apply_dict_object'
+    'apply_dict_object',
+    'Immutable'
 ]
 
 from .dict_object import *
 from .singleton import Singleton
 from .structures import *
 from .typednamedtuple import *
+from .immutable import Immutable