+++ tests/test_coding/test_singleton.py
@@ -1,6 +1,3 @@
-# coding=UTF-8
-from __future__ import print_function, absolute_import, division
-
 import unittest
 
 from satella.coding.structures import Singleton