+++ tests/test_instrumentation/__init__.py
@@ -1,6 +1,3 @@
-# coding=UTF-8
-from __future__ import print_function, absolute_import, division
-
 import logging
 
 logger = logging.getLogger(__name__)