+++ tests/test_instrumentation/test_trace_back.py
@@ -1,7 +1,3 @@
-# coding=UTF-8
-
-from __future__ import print_function, absolute_import, division
-
 import pickle
 import sys
 import unittest